[CPU] Move LZX code to its own files.
